requirements.txt
ag2
aiohappyeyeballs
aiohttp
aiosignal
altair
annotated-types
anyio
appdirs
astrapy
asttokens
asyncer
attrs
autogen
babel
backoff
backrefs
bcrypt
beautifulsoup4
blinker
build
cachetools
cassandra-driver
cassio
certifi
cffi
cfgv
charset-normalizer
chromadb
click
colorama
coloredlogs
crewai
cryptography
dataclasses-json
decorator
deprecation
diskcache
distlib
distro
dnspython
docker
docstring_parser
durationpy
et_xmlfile
executing
fastapi
filelock
flatbuffers
frozenlist
fsspec
geomet
ghp-import
gitdb
GitPython
google-auth
google_search_results
googleapis-common-protos
greenlet
grpcio
h11
h2
hpack
html5lib
httpcore
httptools
httpx
httpx-sse
huggingface-hub
humanfriendly
hyperframe
identify
idna
importlib_metadata
importlib_resources
instructor
ipython
ipython_pygments_lexers
jedi
Jinja2
jiter
json5
json_repair
jsonpatch
jsonpickle
jsonpointer
jsonref
jsonschema
jsonschema-specifications
kubernetes
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langgraph
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
langsmith
litellm
lxml
Markdown
markdown-it-py
MarkupSafe
marshmallow
matplotlib-inline
mdurl
mergedeep
mkdocs
mkdocs-get-deps
mkdocs-material
mkdocs-material-extensions
mmh3
mpmath
multidict
mypy_extensions
narwhals
networkx
nodeenv
numpy
oauthlib
onnxruntime
openai
openpyxl
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
orjson
ormsgpack
overrides
packaging
paginate
pandas
parso
pathspec
pdfminer.six
pdfplumber
pillow
platformdirs
posthog
pre_commit
prompt_toolkit
propcache
protobuf
pure_eval
pyarrow
pyasn1
pyasn1_modules
pybase64
pycparser
pydantic
pydantic-settings
pydantic_core
pydeck
Pygments
PyJWT
pymdown-extensions
pymongo
PyMuPDF
pypdfium2
PyPika
pyproject_hooks
pyreadline3
python-dateutil
python-docx
python-dotenv
pytz
pyvis

PyYAML
pyyaml_env_tag
referencing
regex
requests
requests-oauthlib
requests-toolbelt
rich
rpds-py
rsa
setuptools
shellingham
six
smmap
sniffio
soupsieve
SQLAlchemy
stack-data
starlette
streamlit
sympy
tenacity
termcolor
tiktoken
tokenizers
toml
tomli
tomli_w
tornado
tqdm
traitlets
typer
typing-inspect
typing-inspection
typing_extensions
tzdata
upstash-vector
urllib3
uuid6
uv
uvicorn
virtualenv
watchdog
watchfiles
wcwidth
webencodings
websocket-client
websockets
wheel
xxhash
yarl
zipp
zstandard
pinecone
grpcio
certifi
urllib3
langdetect==1.0.9
deep-translator>=1.11.4

gtts 
transformers 
accelerate
fitz
